[ECP-9072] Implement abstract observer and simplify the observer logic
