+++ Observer/AddPaypalShortcuts.php
@@ -3,7 +3,7 @@
  *
  * Adyen ExpressCheckout Module
  *
- * Copyright (c) 2020 Adyen B.V.
+ * Copyright (c) 2024 Adyen N.V.
  * This file is open source and available under the MIT license.
  * See the LICENSE file for more info.
  *
@@ -11,11 +11,11 @@
  */
 namespace Adyen\ExpressCheckout\Observer;
 
-use Adyen\ExpressCheckout\Block\ApplePay\Shortcut\Button;
+use Adyen\ExpressCheckout\Block\Paypal\Shortcut\Button;
 use Magento\Framework\Event\ObserverInterface;
 
-class AddApplePayShortcuts extends AbstractPaymentMethodShortcuts implements ObserverInterface
+class AddPaypalShortcuts extends AbstractPaymentMethodShortcuts implements ObserverInterface
 {
     const SHORTCUT_BUTTON = Button::class;
-    const PAYMENT_METHOD_VARIANT = Button::APPLE_PAY_VARIANT;
+    const PAYMENT_METHOD_VARIANT = Button::PAYPAL_VARIANT;
 }
